requirements-extras.txt
pandas>=0.23.4
xlsxwriter>=1.1.2
openpyxl>=3.0.0
xarray>=0.13.0
tables>=3.6.1
pyarrow>=0.16.0
msgpack>=1.0.0
msgpack-numpy>=0.4.7
visidata==2.4
